cars_catalog/requirements.txt
flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-restful==0.3.10